Switch to new knncolle libraries with <thread>-based parallelization. (#25)
This uses the new assorthead package that contains the vendored libraries,
instead of the old internal implementations for VP trees and KMKNN. Doing
so allows us to greatly streamline the code in this package.

A major change is that the built indices are no longer R objects. Instead, they
are opaque external pointers into C++ memory - a consequence of deferring to
knncolle's abstraction. These pointers are exposed to developers for re-use
within an R session but cannot be serialized. As this is potentially a breaking
change for some applications, we need to do a major version bump.

We provide pointers to the knncolle::Builder objects for use in C++ code where
the NN algorithm must be specified before accessing the data. This is useful as
downstream developers can just leverage the compiled code in this package when
they want to choose between different algorithms.

We now parallelize using <thread> instead of BiocParallel, which avoids the
overhead of forking or creating a new process.

We replace the last= option with find/queryDistance, which returns the 
distance to the k-th neighbor in a more convenient manner. 

We generalize findKNN() and queryKNN() to allow variable k. Not sure who will
use this, but we did it for distances in findNeighbors(), so might as well.